requirements.txt
Pillow~=7.2.0
PyV4L2Camera~=0.1a2
